--- Pac-Man-2020/Assets/Scripts/Controller2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d75c00f6f7f3e4d76a625e1abe81e937
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
